sysfs: respect lables for sysfs gpio names
In case gpio lines have nimes from platform driver or from DT lines no
longer have names in format of gpioN, but instead the name assigned from
outer source.

This is very hard to predict so it is better to explicitly
specify name (if any) and system offset of gpio pin.

In this case system offset will be used for export, and name will be used
to access gpio pin.

For additional info see drivers/gpio/gpiolib-sysfs.c:gpiod_export.